tools/darktable-gemini-batch/requirements.txt
google-generativeai>=0.8.0
Pillow>=10.0.0
numpy>=1.24.0
rawpy>=0.19.0
imageio>=2.31.0
opencv-python-headless>=4.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
